# packages/credential-showcase-openapi/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Credential Showcase API
  version: 0.1.1
servers:
  - url: https://virtserver.swaggerhub.com/SphereonInt/showcase/0.1.0/v1
    description: SwaggerHub API Auto Mocking
  - url: https://api.example.com/v1
...
security:
  - OAuth2: []
  - OIDC: []
paths:
...
  /workflows/presentations:
    get:
      tags:
        - Presentation Flow
      summary: List all presentation flows
      operationId: listPresentationFlows
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PresentationFlowsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - Presentation Flow
      summary: Create a new presentation flow
      operationId: createPresentationFlow
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PresentationFlowRequest'
        required: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PresentationFlow'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /workflows/presentations/{presentationFlowId}:
    get:
      tags: [ "Presentation Flow" ]
      summary: Retrieve a single presentation flow
      operationId: getPresentationFlow
      parameters:
        - name: presentationFlowId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PresentationFlow'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags: [ "Presentation Flow" ]
      summary: Update a presentation flow (overwrite)
      operationId: updatePresentationFlow
      parameters:
        - name: presentationFlowId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PresentationFlowRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PresentationFlow'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags: [ "Presentation Flow" ]
      summary: Delete a presentation flow
      operationId: deletePresentationFlow
      parameters:
        - name: presentationFlowId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /workflows/presentations/{presentationFlowId}/steps:
    get:
      tags:
        - "Presentation Flow"
      summary: List steps in a presentation flow
      operationId: listPresentationFlowSteps
      parameters:
        - name: presentationFlowId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StepsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags: [ "Presentation Flow" ]
      summary: Add a step to a presentation flow
      operationId: createPresentationFlowStep
      parameters:
        - name: presentationFlowId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StepRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Step'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /workflows/presentations/{presentationFlowId}/steps/{stepId}:
    get:
      tags: [ "Presentation Flow" ]
      summary: Retrieve a specific step in a presentation flow
      operationId: getPresentationFlowStep
      parameters:
        - name: presentationFlowId
          in: path
          required: true
          schema:
            type: string
        - name: stepId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Step'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags: [ "Presentation Flow" ]
      summary: Overwrite a step in a presentation flow
      operationId: updatePresentationFlowStep
      parameters:
        - name: presentationFlowId
          in: path
          required: true
          schema:
            type: string
        - name: stepId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Step'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Step'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags: [ "Presentation Flow" ]
      summary: Delete a step in a presentation flow
      operationId: deletePresentationFlowStep
      parameters:
        - name: presentationFlowId
          in: path
          required: true
          schema:
            type: string
        - name: stepId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /workflows/presentations/{presentationFlowId}/steps/{stepId}/actions:
    get:
      tags: [ "Presentation Flow" ]
      summary: List actions in a step
      operationId: listPresentationFlowStepActions
      parameters:
        - name: presentationFlowId
          in: path
          required: true
          schema:
            type: string
        - name: stepId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StepActionsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags: [ "Presentation Flow" ]
      summary: Create an action in a step
      operationId: createPresentationFlowStepAction
      parameters:
        - name: presentationFlowId
          in: path
          required: true
          schema:
            type: string
        - name: stepId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StepActionRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StepAction'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /workflows/presentations/{presentationFlowId}/steps/{stepId}/actions/{actionId}:
    get:
      tags: [ "Presentation Flow" ]
      summary: Retrieve a single action
      operationId: getPresentationFlowStepAction
      parameters:
        - name: presentationFlowId
          in: path
          required: true
          schema:
            type: string
        - name: stepId
          in: path
          required: true
          schema:
            type: string
        - name: actionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StepAction'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags: [ "Presentation Flow" ]
      summary: Overwrite an action
      operationId: updatePresentationFlowStepAction
      parameters:
        - name: presentationFlowId
          in: path
          required: true
          schema:
            type: string
        - name: stepId
          in: path
          required: true
          schema:
            type: string
        - name: actionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StepActionRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StepAction'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags: [ "Presentation Flow" ]
      summary: Delete an action
      operationId: deletePresentationFlowStepAction
      parameters:
        - name: presentationFlowId
          in: path
          required: true
          schema:
            type: string
        - name: stepId
          in: path
          required: true
          schema:
            type: string
        - name: actionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    OAuth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: <https://auth.example.com/oauth/authorize>
          tokenUrl: <https://auth.example.com/oauth/token>
          scopes:
            editor: Grants editor access
            admin: Grants admin access
            viewer: Grants viewer access
    OIDC:
      type: openIdConnect
      openIdConnectUrl: <https://auth.example.com/.well-known/openid-configuration>
      description: OpenID Connect security scheme for OAuth2 flows
  schemas:
    Asset:
      type: object
      required:
        - id
        - mediaType
        - content
      properties:
        id:
          type: string
          description: Unique identifier for the asset in the database
          example: 123e4567-e89b-12d3-a456-426614174469
        mediaType:
          type: string
          description: Type of asset (e.g., image, document, icon)
          example: image/jpeg
        content:
          type: string
          description: Binary data of the asset
          example: base64 encoded binary data
        fileName:
          type: string
          description: Name of the asset file
          example: asset.jpg
        description:
          type: string
          description: Description of the asset
          example: A beautiful image of a cat
...
    Step:
      type: object
      required:
        - id
        - title
        - description
        - order
        - type
        - actions
      properties:
        id:
          type: string
          description: Unique identifier for the step
          example: 123e4567-e89b-12d3-a456-434314174000
        title:
          type: string
          description: Title of the step
          example: Verify Identity
        description:
          type: string
          description: Detailed description of the step
          example: Verify the user's identity
        order:
          type: integer
          description: Order of the step in the workflow
          minimum: 0
          example: 1
        type:
          $ref: '#/components/schemas/StepType'
        subFlow:
          type: string
          description: Optional sub-workflow for this step
          example: 123e4567-e89b-12d3-a456-434314174000
        actions:
          type: array
          items:
            $ref: '#/components/schemas/StepAction'
        asset:
          $ref: '#/components/schemas/Asset'
    StepRequest:
      type: object
      required:
        - title
        - description
        - order
        - type
        - actions
      properties:
        title:
          type: string
          description: Title of the step
          example: Verify Identity
        description:
          type: string
          description: Detailed description of the step
          example: Verify the user's identity
        order:
          type: integer
          description: Order of the step in the workflow
          minimum: 0
          example: 1
        type:
          $ref: '#/components/schemas/StepType'
        subFlow:
          type: string
          description: Optional sub-workflow for this step
          example: 123e4567-e89b-12d3-a456-434314174000
        actions:
          type: array
          description: List of actions associated with this step
          items:
            $ref: '#/components/schemas/StepActionRequest'
        asset:
          type: string
          description: Asset referenced by this step
          example: 456e4567-e89b-12d3-a456-426614174000
    StepsResponse:
      properties:
        steps:
          type: array
          items:
            $ref: '#/components/schemas/Step'
...
    StepType:
      type: string
      enum:
        - HUMAN_TASK
        - SERVICE
        - WORKFLOW
      description: Type of step in the workflow
      example: HUMAN_TASK
    StepAction:
      type: object
      required:
        - id
        - actionType
        - title
        - text
      properties:
        id:
          type: string
          description: Unique identifier for the action
          example: 123e4567-ef2d-12d3-abcd-426614174456
        actionType:
          type: string
          description: Type of action
          example: EXAMPLE_ACTION
        title:
          type: string
          description: Title of the action
          example: Connect Wallet
        text:
          type: string
          description: Descriptive text for the action
          example: Connect your wallet to continue
      discriminator:
        propertyName: actionType
        mapping:
          ARIES_OOB: '#/components/schemas/AriesOOBAction'
    StepActionRequest:
      type: object
      required:
        - actionType
        - title
        - text
      properties:
        title:
          type: string
          description: Title of the action
          example: Connect Wallet
        text:
          type: string
          description: Descriptive text for the action
          example: Connect your wallet to continue
        actionType:
          type: string
          example: HUMAN_TASK
      discriminator:
        propertyName: actionType
        mapping:
          ARIES_OOB: '#/components/schemas/AriesOOBAction'
    StepActionsResponse:
      properties:
        actions:
          type: array
          items:
            $ref: '#/components/schemas/StepAction'
...
    AriesOOBAction:
      allOf:
        - $ref: '#/components/schemas/StepAction'
        - type: object
          required:
            - proofRequest
          properties:
            proofRequest:
              $ref: '#/components/schemas/AriesProofRequest'
    AriesProofRequest:
      type: object
      properties:
        attributes:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/AriesRequestCredentialAttributes'
          description: Map of attribute groups requested in the proof
          example:
            name:
              value: John Doe
        predicates:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/AriesRequestCredentialPredicates'
          description: Map of predicate groups requested in the proof
          example:
            age:
              value: 30
...
    Persona:
      type: object
      required:
        - name
        - role
      properties:
        id:
          type: string
          description: Unique identifier for the persona
          example: 123e4567-e89b-12d3-a456-426614174456
        name:
          type: string
          description: Name of the persona
          example: John Doe
        role:
          type: string
          description: Role of the persona in the system
          example: Verifier
        description:
          type: string
          description: Detailed description of the persona
          example: John Doe is a verifier for the system
        headshotImage:
          $ref: '#/components/schemas/Asset'
        bodyImage:
          $ref: '#/components/schemas/Asset'
...
    Workflow:
      type: object
      required:
        - name
        - type
      properties:
        id:
          type: string
          description: Unique identifier for the workflow
          example: 789e4567-e89b-12d3-a456-434314174123
        name:
          type: string
          description: Name of the workflow
          example: Credential Issuance
        description:
          type: string
          description: Detailed description of the workflow
          example: This workflow issues credentials to users
        type:
          type: string
          enum: [ ISSUANCE, PRESENTATION ]
          description: Type of workflow
          example: ISSUANCE
        steps:
          type: array
          description: Ordered list of steps in the workflow
          items:
            $ref: '#/components/schemas/Step'
        personas:
          type: array
          description: References to personas used in this workflow
          items:
            $ref: '#/components/schemas/Persona'
    WorkflowRequest:
      type: object
      required:
        - name
        - description
      properties:
        name:
          type: string
          description: Name of the workflow
          example: Credential Issuance
        description:
          type: string
          description: Detailed description of the workflow
          example: This workflow issues credentials to users
        steps:
          type: array
          description: Ordered list of steps in the workflow
          items:
            $ref: '#/components/schemas/StepRequest'
        personas:
          type: array
          description: References to personas used in this workflow
          items:
            type: string
            example: 123e4567-e89b-12d3-a456-426614174456
...
    PresentationFlowsResponse:
      type: object
      properties:
        presentationFlows:
          type: array
          items:
            $ref: '#/components/schemas/PresentationFlow'
...
    PresentationFlow:
      type: object
      allOf:
        - $ref: '#/components/schemas/Workflow'
        - type: object
          required:
            - relyingParty
          properties:
            relyingParty:
              $ref: '#/components/schemas/RelyingParty'
      description: Workflow specialization for credential presentation
    PresentationFlowRequest:
      allOf:
        - $ref: '#/components/schemas/WorkflowRequest'
        - type: object
          required:
            - relyingParty
          properties:
            relyingParty:
              type: string
              example: "313e4567-e89b-12d3-a456-426614174469"
...
    RelyingPartyType:
      type: string
      enum:
        - ARIES
      description: Types of supported relying parties
...
    RelyingParty:
      type: object
      required:
        - id
        - name
        - description
        - type
        - credentialDefinitions
      properties:
        id:
          type: string
          description: Unique identifier for the relying party
          example: 76543210-e89b-12d3-a456-426614174469
        name:
          type: string
          description: Name of the relying party
          example: Relying Party Name
        description:
          type: string
          description: Detailed description of the relying party
          example: This relying party verifies credentials from issuers
        type:
          $ref: '#/components/schemas/RelyingPartyType'
        organization:
          type: string
          description: Organization the relying party belongs to
          example: Acme Corporation
        logo:
          $ref: '#/components/schemas/Asset'
        credentialDefinitions:
          type: array
          items:
            $ref: '#/components/schemas/CredentialDefinition'
...
    CredentialType:
      type: string
      enum:
        - ANONCRED
      description: Types of supported credentials
...
    CredentialDefinition:
      type: object
      required:
        - id
        - name
        - version
        - type
        - schemaId
        - icon
        - representations
      properties:
        id:
          type: string
          description: Unique identifier for the credential definition
          example: 123e4567-e89b-12d3-a456-426614174123
        name:
          type: string
          description: Name of the credential definition
          example: Credential Definition Name
        issuerId:
          type: string
          description: Unique identifier of the issuer who created this schema
          example: 123e4567-e89b-12d3-a456-426614174123
        schemaId:
          type: string
          description:  Identifier of the associated credential schema
          example:  123e4567-e89b-12d3-a456-426614174123
        version:
          type: string
          description: Version of the credential definition
          example: '1.0'
        type:
          $ref: '#/components/schemas/CredentialType'
        representations:
          type: array
          items:
            oneOf:
              - $ref: '#/components/schemas/CredentialRepresentation'
              - $ref: '#/components/schemas/OCARepresentation'
        revocation:
          oneOf:
            - $ref: '#/components/schemas/RevocationInfo'
            - $ref: '#/components/schemas/AnonCredRevocation'
        icon:
          $ref: '#/components/schemas/Asset'
...
    CredentialRepresentation:
      type: object
      required:
        - id
      properties:
        id:
          type: string
          description: Unique identifier for the credential representation
          example: 123e4567-e89b-12d3-abcd-426614174456
...
    AnonCredRevocation:
      allOf:
        - $ref: '#/components/schemas/RevocationInfo'
        - type: object
          required:
            - registryId
          properties:
            registryId:
              type: string
              description: Identifier of the revocation registry for AnonCreds
              example: 123e4567-e89b-12d3-a456-426614174123
      description: Anonymous Credentials specific revocation information
...
  responses:
    'BadRequest':
      description: Bad Request
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: 'Invalid request data'
    'Unauthorized':
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: 'Invalid authentication credentials'
    'Forbidden':
      description: Forbidden
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: 'Insufficient permissions to access resource'
    'NotFound':
      description: Not Found
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: 'Resource not found'
    'InternalServerError':
      description: Internal Server Error
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: 'An unexpected error occurred'